Remove duplicated joboffers
